# Backend/Microservices/Gestion_User/src/main/java/org/esprit/gestion_user/Service/AuthenticationResponse.java
package org.esprit.gestion_user.Service;

import lombok.Builder;
import lombok.Getter;
import lombok.Setter;

@Getter
@Setter
@Builder
public class AuthenticationResponse {
    private String token;
}